BUG: Ensure consistent ordering of amounts, prices & multipliers
Previously, we have assumed that the `amounts` and `last_sale_prices`
lists have the same order as the `value_multipliers`. This is not
correct, since to populate the `amounts` and `last_sale_prices` lists
we iterate over a `dict` (self.positions) which means we don't know
the order to the keys/values in advance. Our `value_multipliers` however
are stored in an `OrderedDict`.

To address this issue, we make sure that `self.positions` subclasses
`OrderedDict`.